Moving the input to Qiskit Aqua Finance
