+++ qiskit/aqua/input/finance/algorithminput.py
