+++ qiskit/aqua/input/finance/exchangedata/README.md
