+++ qiskit/aqua/input/finance/exchangedata/__init__.py
